Changed Department from String to Enum (DepartmentType), find and save department in department repository, id service scope changed to ProtoType, and displaying department id and description in htmlt, included the LoggingDepartmentsAspect

+++ src/main/java/com/schoolar/sb/api/PersonRestController.java
@@ -36,7 +36,7 @@ public class PersonRestController {
     }
 
     @GetMapping( "/{personId}" )
-    public ResponseEntity<PersonDto> getPersonById( @PathVariable Integer personId ) {
+    public ResponseEntity<PersonDto> getPersonById( @PathVariable Long personId ) {
 
         var person = personRepository.findByPersonId( personId );
         var personDto = personMapper.toTdo( person );
